Better styling for the counter. Reorganizing the components.
