+++ src/screens/HomeScreen.tsx
@@ -1,14 +1,11 @@
 import { Container } from 'native-base'
 import React from 'react'
-import { SafeAreaView } from 'react-native'
 import { Home } from '../components/Home'
 
 export default function HomeScreen() {
   return (
-    <Container>
-      <SafeAreaView>
-        <Home />
-      </SafeAreaView>
+    <Container style={{ paddingHorizontal: 10 }}>
+      <Home />
     </Container>
   )
 }
